pyproject.toml
[tool.poetry]
name = "youtube-mcp-server-with-fastmcp"
version = "0.1.0"
description = ""
authors = ["Gisela Torres"]
readme = "README.md"
packages = [{include = "src"}]
include = ["assets/**/*"]

[tool.poetry.dependencies]
python = "^3.10"
fastmcp = "2.14.2"
google-api-python-client = "^2.187.0"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
